+++ src/main/java/roomescape/exception/NotFoundReservationException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class NotFoundReservationException extends RuntimeException {
